function 2/requirements.txt
# Function dependencies, for example:
pandas
google-cloud-storage
gcsfs